remove Cards directory and refactor tests that were affected by this
