--- app-web/src/components/Cards/Card/Card.module.css
@@ -1,15 +1,0 @@
-.Card {
-    width: 250px;
-    height: 350px;
-    background-color: #fff;
-    /* padding: 14px 17px 32px; */
-    position: relative;
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: space-between;
-    color: #3b3b3b;
-    flex-shrink: 0;
-    margin: 10px;
-    box-shadow: 0 1px 2px 2px #ccc;
-    transition: box-shadow 0.25s ease-out, transform .25s ease-out;
-}
