--- app-web/src/components/Cards/Cards.module.css
@@ -1,40 +1,0 @@
-.TopicContainer h1 {
-    flex-grow: 1;
-    font-size: 24px;
-    border-bottom: 0;
-    font-weight: 400;
-    color: #494949;
-    margin-left: 10px;
-}
-
-.TopicContainer p {
-    margin: 9px 0 15px 10px;
-}
-
-.TopicContainer a {
-    color: #242424;
-    font-size: 28px;
-}
-
-.ToggleButton {
-    text-align: left;
-    font-size: 17px;
-}
-
-.LargeView {
-    display: none;
-}
-
-.MobileView {
-    display: block;
-}
-
-@media screen and (min-width: 570px) {
-    .MobileView {
-        display: none;
-    }
-
-    .LargeView {
-        display: block;
-    }
-}
